Separate out the provider class
